Checkpoint 01-08-2026 02:22

+++ backend/CommentService/src/CommentService.Application/UseCases/Comments/RemoveByTypeComment.cs
@@ -1,18 +1,19 @@
-using CommentService.Application.UseCases.Comments.Interfaces;
 using CommentService.Application.Interfaces;
+using CommentService.Application.UseCases.Comments.Interfaces;
+using CommentService.Domain.Enums;
 using CommentService.Application.Caching.Comment;
 using CommentService.Application.Validators.Interfaces;
 using CommentService.Domain.Entities;
 using CommentService.Application.Exceptions;
 namespace CommentService.Application.UseCases.Comments
 {
-    public class RemoveByModeratorComment : IRemoveByModeratorComment
+    public class RemoveByTypeComment : IRemoveByTypeComment
     {
         private readonly ICommentServices commentServices;
         private readonly ICommentCacheServices commentCacheServices;
         private readonly IRabbitMQProducer rabbitMQProducer;
         private readonly ICommentValidationService commentValidationService;
-        public RemoveByModeratorComment(
+        public RemoveByTypeComment(
             ICommentServices _commentServices,
             ICommentCacheServices _commentCacheServices,
             IRabbitMQProducer _rabbitMQProducer,
@@ -24,11 +25,22 @@ namespace CommentService.Application.UseCases.Comments
             this.rabbitMQProducer = _rabbitMQProducer;
             this.commentValidationService = _commentValidationService;
         }
-        public async Task ExecuteAsync(Guid authenticatedUserId, Guid commentId)
+        public async Task ExecuteAsync(Guid authenticatedUserId, Guid commentId, CommentDeletionType deletionType)
         {
             await this.commentValidationService.ValidateUserExists(authenticatedUserId);
             var comment = await GetComment(commentId);
-            comment.DeleteByModerator();
+            if(comment.UserProjection.UserId != authenticatedUserId)
+                throw new ValidationException("Você não pode remover este comentário.");
+            switch (deletionType)
+            {
+                case CommentDeletionType.User:
+                    comment.DeleteByUser();
+                    break;
+                case CommentDeletionType.Moderator:
+                    comment.DeleteByModerator();
+                    break;
+                
+            }
             await this.commentServices.Update(comment);
             var type = comment.Type.ToString();
             await this.rabbitMQProducer.Publish($"{type}CommentDeleted",
